utilized context api to create seperate component for the bottom left portion of dashboard that controls pruning buttons/commands
